code/requirements.txt
selenium
django
psycopg2
django-filter
django-mathfilters
django-crispy-forms
django-widget-tweaks
nltk
syllables
googletrans
coverage
django-nose